requirements.txt
aiogram>=3
openai
python-dotenv
